实现用户聊天，但是思路错误，本质还是短连接。只是加了一个自动刷新机制，每隔一秒就通过短连接向服务端发去获取更新数据，然后更新聊天界面UI

+++ Client/src/controller/CreateGroupChat.java
@@ -80,7 +80,7 @@ public class CreateGroupChat {
         System.out.println();
         System.out.println("雪花算法生成的群ID："+name_id);
 
-        Group newGroup = new Group(name_id,groupName, groupDesc, "/image/默认头像.png", currentUser.getId());
+        Group newGroup = new Group(name_id,groupName, groupDesc, "/image/默认头像.png", (int) currentUser.getId());
         Gson gson = new Gson();
         String groupJson = gson.toJson(newGroup);
